revert to Last6TransactionFinder

+++ STAGGI-Budget-API/Services/Interfaces/IBudgetService.cs
@@ -9,7 +9,7 @@ namespace STAGGI_Budget_API.Services.Interfaces
     {
         
         public Result<List<BudgetDTO>> GetAllByEmail(string email);
-        public Result<BudgetDTO> GetById(long id);
+        public Result<BudgetDTO> GetById(int id);
         public Result<string> CreateBudget(RequestBudgetDTO budgetDTO, string email);
         //public Result<string> DeleteBudget(long budgetId);
         public Result<BudgetDTO> UpdateBudget(int budgetId, RequestBudgetDTO updatedBudgetDTO, string email);
